reshape the changes
