+++ internal/orchestrator/orchestrator_test.go
@@ -640,7 +640,7 @@ models:
 	modelIndex, err := modelsindex.Load(platform.GetPlatform(nil), cfg.AssetDir(), cfg.ModelsDir(), cfg.CustomModelsDir(), nil, config.Configuration{})
 	require.NoError(t, err)
 
-	env := AppEnvironment(t.Context(), appDesc, bricksIndex, modelIndex, platform.Platform{}, cfg).All()
+	env := appEnvironment(t.Context(), appDesc, bricksIndex, modelIndex, platform.Platform{})
 	require.Equal(t, cfg.AppsDir().Join("app1").String(), env["APP_HOME"])
 	require.Equal(t, cfg.ModelsDir().String(), env["MODELS_PATH"])
 	require.Equal(t, "/models/ootb/ei/yolo-x-nano.eim", env["EI_OBJ_DETECTION_MODEL"])
@@ -721,7 +721,7 @@ models:
 	modelIndex, err := modelsindex.Load(platform.GetPlatform(nil), cfg.AssetDir(), cfg.ModelsDir(), cfg.CustomModelsDir(), nil, config.Configuration{})
 	require.NoError(t, err)
 
-	env := AppEnvironment(t.Context(), appDesc, bricksIndex, modelIndex, platform.Platform{}, cfg).All()
+	env := appEnvironment(t.Context(), appDesc, bricksIndex, modelIndex, platform.Platform{})
 	require.Equal(t, cfg.AppsDir().Join("app1").String(), env["APP_HOME"])
 	require.Equal(t, "/home/arduino/.arduino-bricks/models/face-det.eim", env["EI_OBJ_DETECTION_MODEL"])
 	require.Equal(t, "/home/arduino/.arduino-bricks/models", env["CUSTOM_MODEL_PATH"])
@@ -804,7 +804,7 @@ models:
 	modelIndex, err := modelsindex.Load(platform.GetPlatform(nil), cfg.AssetDir(), cfg.ModelsDir(), cfg.CustomModelsDir(), nil, config.Configuration{})
 	require.NoError(t, err)
 
-	env := AppEnvironment(t.Context(), appDesc, bricksIndex, modelIndex, platform.Platform{}, cfg).All()
+	env := appEnvironment(t.Context(), appDesc, bricksIndex, modelIndex, platform.Platform{})
 	require.Equal(t, "/models/path/obj.eim", env["EI_OBJ_DETECTION_MODEL"])
 	require.Equal(t, "/models/path/video.eim", env["EI_V_OBJ_DETECTION_MODEL"])
 	require.Equal(t, "/default/video/value", env["MY_VIDEO_ENV"])
